patch from Olivier Coursiere for better error output in ping change the size parameter type of several functions in sys/socket.h to match POSIX compat libs and legacy headers keep the original R5 type (though I make a change for this)
git-svn-id: file:///srv/svn/repos/haiku/haiku/trunk@17911 a95241bf-73f2-0310-859d-f6bbb57e9c96

+++ src/kits/network/compat/libnet/res_send.c
@@ -560,7 +560,7 @@ read_len:
 			struct timeval timeout;
 			fd_set *dsmaskp;
 			struct sockaddr_storage from;
-			size_t fromlen;
+			int fromlen;
 
 			if ((s < 0) || vc || (af != nsap->sa_family)) {
 				if (vc)
